+++ Changes.md
@@ -2,6 +2,7 @@
 - Support type definitions in first-class module types.
 - Handle correctly first-class module types with nested modules inside.
 - Add support for first-class module types with type equations.
+- Add support for functions with first-class module types and type equations.
  
 # 2.11.0
 - Fix: don't export types defined in shims as opaque.